#15 fix : gateway - service 분리
- GameService가 유저의 위치를 기억하도록 변경
- 중복함수 getBombUser 삭제
