+++ src/status/status.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, Logger } from "@nestjs/common";
 
 @Injectable()
 export class StatusBombGameService {
@@ -6,6 +6,17 @@ export class StatusBombGameService {
   private bombUserList: string[] = [];
   private playUserCount: number;
   private BOMB_USER_PERCENT: number = 0.2;
+  private BOMB_RADIUS: number = 40;
+  private TAG_HOLD_DURATION_MS: number = 1500;
+
+  private bombGameRoomPosition: Map<string, { room: string, x: string, y: string, isStun: number }> = new Map();
+  private temporarilyExcludedUsers = new Map<string, NodeJS.Timeout>();
+
+  private logger: Logger = new Logger("BombGameService");
+
+  setBombGamePlayerRoomPosition(playId,{room, x, y, isStun}){
+    this.bombGameRoomPosition.set(playId,{room:room,x:x,y:x,isStun:isStun})
+  }
 
   setPlayGameUser(playList: Set<string>) {
     // 기존 관리되고 있던 유저 초기화
@@ -18,17 +29,12 @@ export class StatusBombGameService {
 
     return this.playGameUser;
   }
-
-  getBombUsers(): string[] {
-    return this.bombUserList;
-  }
-
+  
   getNewBombUsers(): string[] {
     this.bombUserList = this.selectRandomBombUsers();
     return this.bombUserList;
   }
-
-
+  
   getPlayGameUserList(): string[] {
     return Array.from(this.playGameUser);
   }
@@ -56,6 +62,47 @@ export class StatusBombGameService {
       return this.getPlayGameUserList();
     }
     return null;
+  }
+
+  checkOverlappingUser(): boolean{
+    const updatedBombUserList: string[] = [];
+    let updated = false;
+
+    this.bombUserList.forEach(bombUserId => {
+      const bombUserPosition = this.bombGameRoomPosition.get(bombUserId);
+      if (bombUserPosition) {
+        const overlappingUser = Array.from(this.bombGameRoomPosition.entries()).find(([userId, position]) => {
+          if (userId !== bombUserId && this.bombGameRoomPosition.has(userId) && position.room === bombUserPosition.room && !this.temporarilyExcludedUsers.has(userId)) {
+            const distance = Math.sqrt(Math.pow(parseFloat(bombUserPosition.x) - parseFloat(position.x), 2) + Math.pow(parseFloat(bombUserPosition.y) - parseFloat(position.y), 2));
+            return distance <= this.BOMB_RADIUS;
+          }
+          return false;
+        });
+
+        if (overlappingUser) {
+          updatedBombUserList.push(overlappingUser[0]);
+          updated = true;
+          // 새롭게 술래가 된 유저를 일정 시간 동안 술래 상태에서 유지
+          const timeout = setTimeout(() => {
+            this.temporarilyExcludedUsers.delete(overlappingUser[0]);
+            this.temporarilyExcludedUsers.delete(bombUserId);
+          }, this.TAG_HOLD_DURATION_MS); // 술래 상태를 유지
+
+          this.temporarilyExcludedUsers.set(overlappingUser[0], timeout);
+          this.temporarilyExcludedUsers.set(bombUserId, timeout);
+        } else {
+          updatedBombUserList.push(bombUserId);
+        }
+      }
+    });
+    this.bombUserList=updatedBombUserList
+    this.logger.debug(`Updated bombUserList: ${this.bombUserList}`);
+    return updated
+  }
+
+
+  getBombUserList():string[]{
+      return this.bombUserList;
   }
 
   private selectRandomBombUsers(): string[] {
